requirements.txt
Flask==1.0.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
stemming==1.0
google-cloud-translate==1.3.3
googleapis-common-protos==1.5.9
numpy==1.15.4
numpydoc==0.8.0
sklearn==0.0
cloudpickle==0.6.1
python-daemon==2.1.2
python-dateutil==2.7.5
python-pptx==0.6.5
GitPython==2.1.11
ipython==7.2.0
ipython-genutils==0.2.0
gunicorn==19.9.0
h5py==2.9.0